requirements.txt
streamlit>=1.32.0
pandas>=2.2.0
PyPDF2>=3.0.0
python-docx>=1.1.0
spacy>=3.7.0
scikit-learn>=1.4.0
together>=0.2.10
pytesseract>=0.3.10
pdf2image>=1.17.0
python-dotenv>=1.0.0
numpy>=1.26.0 